Move kapteyn_celestial.py to extern and fix coveragerc
