+++ pyregion/tests/coveragerc
@@ -11,6 +11,7 @@ omit =
    {packagename}/*/tests/*
    {packagename}/*/*/tests/*
    {packagename}/version*
+   {packagename}/extern/*
 
 [report]
 exclude_lines =
